catalog/requirements.txt
asgiref==3.8.1
Django==3.2
django-utils-six==2.0
djangorestframework==3.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
